requirements.txt
llama-index==0.10.0
llama-index-vector-stores-faiss==0.1.2
llama-index-llms-groq==0.1.3
llama-index-embeddings-huggingface==0.1.4
faiss-cpu==1.7.4
python-dotenv==1.0.0
torch==2.1.0
transformers==4.36.0
pypdf
pymupdf
pdfplumber
